Move legacy inference utils
